wpt_babel/content-security-policy/script-src-attr-elem/script-src-attr-blocked-src-allowed.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="script-src-attr 'none';
    script-src 'unsafe-inline' 'self';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <script>
   "use strict";

var t = async_test("Should fire a security policy violation event");
window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
  assert_equals(e.violatedDirective, 'script-src-attr');
  assert_equals(e.blockedURI, 'inline');
}));
  </script>
  <img onload="t.unreached_func('Should not have executed the inline handler')" src="../support/pass.png"/>
 </body>
</html>
